Add files via upload
Welcome to the Netflix Data Analysis project! This repository contains a collection of scripts and notebooks that demonstrate the use of NumPy and pandas to analyze the Netflix dataset. This project is aimed at providing insights into Netflix's movie and TV show offerings through various data analysis techniques.This project uses the Netflix dataset to showcase the capabilities of NumPy and pandas in data cleaning, manipulation, and analysis. It includes examples of data preprocessing, exploration, and visualization to uncover interesting trends and patterns in Netflix's catalog.